### 2024-02-17 283d113
added sequential testing demo, bugfix sequential testing, add to API reference, updated docs for all pvalues and add pvalue range notices
