+++ contextualized/analysis/pvals.py
@@ -15,6 +15,21 @@ from contextualized.analysis.effects import (
 from contextualized.easy.wrappers import SKLearnWrapper
 
 
+def get_pval_range(num_bootstraps: int) -> list:
+    """
+    Get the range of possible p-values based on the number of bootstraps.
+
+    Args:
+        num_bootstraps (int): The number of bootstraps.
+
+    Returns:
+        list: The minimum and maximum possible p-values.
+    """
+    min_pval = 1 / (num_bootstraps + 1)
+    max_pval = num_bootstraps / (num_bootstraps + 1)
+    return [min_pval, max_pval]
+
+
 def calc_pval_bootstraps_one_sided(estimates, thresh=0, laplace_smoothing=1):
     """
     Calculate p-values from bootstrapped estimates.
@@ -52,25 +67,33 @@ def calc_pval_bootstraps_one_sided_mean(estimates, laplace_smoothing=1):
 
 
 def calc_homogeneous_context_effects_pvals(
-    model: SKLearnWrapper, C: np.ndarray, **kwargs
+    model: SKLearnWrapper, C: np.ndarray, verbose: bool = True, **kwargs
 ) -> np.ndarray:
     """
-    Calculate p-values for the effects of context.
+    Calculate p-values for the effects of context directly on the outcome.
 
     Args:
         model (SKLearnWrapper): Model to analyze.
         C (np.ndarray): Contexts to analyze.
+        verbose (bool): Whether to print the range of possible p-values.
 
     Returns:
         np.ndarray: P-values of shape (n_contexts, n_outcomes) testing whether the
             sign of the direct effect of context on outcomes is consistent across bootstraps.
     """
-    _, effects = get_homogeneous_context_effects(model, C, **kwargs)
-    # effects.shape: (n_contexts, n_bootstraps, n_context_vals, n_outcomes)
-    if len(effects.shape) < 4:
+    if model.n_bootstraps < 2:
         print("P values are not well defined without multiple bootstrap samples.")
         return None
-
+    if verbose:
+        print(
+            "########################################################################################\n"
+            f"This model contains {model.n_bootstraps} bootstraps.\n"
+            f"The range of possible p-values is {get_pval_range(model.n_bootstraps)}.\n"
+            f"To get lower p-values, increase the model's n_bootstraps.\n"
+            "########################################################################################"
+        )
+    _, effects = get_homogeneous_context_effects(model, C, **kwargs)
+    # effects.shape: (n_contexts, n_bootstraps, n_context_vals, n_outcomes)
     diffs = effects[:, :, -1] - effects[:, :, 0]  # Test whether the sign is consistent
     pvals = np.array(
         [
@@ -90,7 +113,7 @@ def calc_homogeneous_context_effects_pvals(
 
 
 def calc_homogeneous_predictor_effects_pvals(
-    model: SKLearnWrapper, C: np.ndarray, **kwargs
+    model: SKLearnWrapper, C: np.ndarray, verbose: bool = True, **kwargs
 ) -> np.ndarray:
     """
     Calculate p-values for the context-invariant effects of predictors.
@@ -98,16 +121,25 @@ def calc_homogeneous_predictor_effects_pvals(
     Args:
         model (SKLearnWrapper): Model to analyze.
         C (np.ndarray): Contexts to analyze.
+        verbose (bool): Whether to print the range of possible p-values.
 
     Returns:
         np.ndarray: P-values of shape (n_predictors, n_outcomes) testing whether the
             sign of the context-invariant predictor effects are consistent across bootstraps.
     """
-    _, effects = get_homogeneous_predictor_effects(model, C, **kwargs)
[…]
