+++ docs/source/analysis.rst
@@ -13,7 +13,7 @@ All functions can be loaded directly from the module, e.g. ``from contextualized
     pvals.calc_homogeneous_predictor_effects_pvals
     pvals.calc_heterogeneous_predictor_effects_pvals
     pvals.test_each_context
-    pvals.get_pval_range
+    pvals.get_possible_pvals
     accuracy_split.print_acc_by_covars
     embeddings.plot_lowdim_rep
     embeddings.plot_embedding_for_all_covars
@@ -25,7 +25,7 @@ All functions can be loaded directly from the module, e.g. ``from contextualized
 .. autofunction:: contextualized.analysis.pvals.calc_homogeneous_predictor_effects_pvals
 .. autofunction:: contextualized.analysis.pvals.calc_heterogeneous_predictor_effects_pvals
 .. autofunction:: contextualized.analysis.pvals.test_each_context
-.. autofunction:: contextualized.analysis.pvals.get_pval_range
+.. autofunction:: contextualized.analysis.pvals.get_possible_pvals
 .. autofunction:: contextualized.analysis.accuracy_split.print_acc_by_covars
 .. autofunction:: contextualized.analysis.embeddings.plot_lowdim_rep
 .. autofunction:: contextualized.analysis.embeddings.plot_embedding_for_all_covars
