+++ contextualized/analysis/__init__.py
@@ -17,5 +17,5 @@ from contextualized.analysis.pvals import (
     calc_homogeneous_predictor_effects_pvals,
     calc_heterogeneous_predictor_effects_pvals,
     test_each_context,
-    get_pval_range,
+    get_possible_pvals,
 )
