### 2024-03-12 #6469 @editorialbot
[REVIEW]: Contextualized: Heterogeneous Modeling Toolbox
**Submitting author:** @cnellington (<a href="http://orcid.org/0000-0001-7029-8023">Caleb Ellington</a>)
**Repository:** https://github.com/cnellington/Contextualized
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.2.8
