+++ joss/paper.md
@@ -67,18 +67,18 @@ With the flexibility of context-dependent parameters, each context-specific mode
 
 `Contextualized` opens up new avenues for quantitative analysis of complex and heterogeneous data, and simplifies the process of transforming this data into results. In particular, `Contextualized`:
 
-1. **Unifies Modeling Frameworks:** `Contextualized` unifies modeling approaches for both homogeneous and heterogeneous data, including population, sub-population, (latent) mixture, cluster-based, time-varying, and varying-coefficient models [@hastie1993varying]. 
+1. **Unifies Modeling Frameworks:** `Contextualized` unifies modeling approaches for both homogeneous and heterogeneous data, including population, sub-population, (latent) mixture, cluster-based, time-varying, and varying-coefficient models [@hastie_varying-coefficient_1993]. 
 Additionally, `Contextualized` naturally falls back to more traditional modeling frameworks when complex heterogeneity is not present.
 Not only is this convenient, but it limits the number of modeling decisions and validation tests required by users, reducing the risk of misspecification and false discoveries [@lengerich_contextualized_2023].
 2. **Models High-resolution Heterogeneity:** Contextualized models adapt to the context of each sample by using a context encoder, naturally accounting for high-dimensional, continuous, and fine-grained variation between samples [@ellington_contextualized_2023].
-3. **Quantifies Heterogeneity in Data:** Context-specific models quantify the randomness and structure of the systems underlying each data point, and variation in context-specific model parameters quantifies the heterogeneity between data points [@deuschel_contextualized_2023].
+3. **Quantifies Heterogeneity in Data:** Context-specific models quantify the randomness and structure of the systems underlying each data point, and variation in context-specific model parameters quantifies the heterogeneity between data points [@deuschel_contextualized_2023; @al-shedivat_personalized_2018].
 `Contextualized` provides tools to analyze, test, and validate contextualized models, unlocking new studies of structured heterogeneity.
 4. **Interpolates and Extrapolates to Unseen Contexts:** By using context encoders to translate between contextual information and model parameters, `Contextualized` learns meta-relationships between metadata and data. At test time, `Contextualized` can adapt to contexts which were never observed in the training data [@ellington_contextualized_2023].
 5. **Analyzes Latent Processes:** By associating structured models with each sample, `Contextualized` enables analysis of samples with latent processes.
-These latent processes can be inferred from patterns in context-specific models, and can be used to identify latent subgroups, latent trajectories, and latent features that explain heterogeneity [@lengerich_discriminative_2020].
+These latent processes can be inferred from patterns in context-specific models, and can be used to identify latent subgroups, latent trajectories, and latent features that explain heterogeneity [@lengerich_discriminative_2022].
 6. **Provides Direct Interpretability:** `Contextualized` estimates and analyzes context-specific statistical models. 
 These statistical models are mathematically-constrained such that each parameter has specific meaning, permitting direct interpretation and immediate results [@lengerich_automated_2022].
-7. **Incorporates Multi-modal Data:** Context is a general and flexible concept, and context-encoders can be used to instill any type of contextual information into contextualized models, including images, text, tabular data, and more [@lengerich_discriminative_2020; @lengerich_notmad_2021; @al-shedivat_contextual_2020; @stoica_contextual_2020].
+7. **Incorporates Multi-modal Data:** Context is a general and flexible concept, and context-encoders can be used to instill any type of contextual information into contextualized models, including images, text, tabular data, and more [@lengerich_discriminative_2022; @lengerich_notmad_2021; @al-shedivat_contextual_2020; @stoica_contextual_2020].
[…]
