+++ joss/paper.bib
@@ -1,90 +1,3 @@
-@article{hastie1993varying,
-  title={Varying-coefficient models},
-  author={Hastie, Trevor and Tibshirani, Robert},
-  journal={Journal of the Royal Statistical Society: Series B (Methodological)},
-  volume={55},
-  number={4},
-  pages={757--779},
-  year={1993},
-  publisher={Wiley Online Library}
-}
-
-@misc{deuschel_contextualized_2023,
-	title = {Contextualized {Policy} {Recovery}: {Modeling} and {Interpreting} {Medical} {Decisions} with {Adaptive} {Imitation} {Learning}},
-	shorttitle = {Contextualized {Policy} {Recovery}},
-	url = {http://arxiv.org/abs/2310.07918},
-	doi = {10.48550/arXiv.2310.07918},
-	abstract = {Interpretable policy learning seeks to estimate intelligible decision policies from observed actions; however, existing models fall short by forcing a tradeoff between accuracy and interpretability. This tradeoff limits data-driven interpretations of human decision-making process. e.g. to audit medical decisions for biases and suboptimal practices, we require models of decision processes which provide concise descriptions of complex behaviors. Fundamentally, existing approaches are burdened by this tradeoff because they represent the underlying decision process as a universal policy, when in fact human decisions are dynamic and can change drastically with contextual information. Thus, we propose Contextualized Policy Recovery (CPR), which re-frames the problem of modeling complex decision processes as a multi-task learning problem in which complex decision policies are comprised of context-specific policies. CPR models each context-specific policy as a linear observation-to-action mapping, and generates new decision models \${\textbackslash}textit\{on-demand\}\$ as contexts are updated with new observations. CPR is compatible with fully offline and partially observable decision environments, and can be tailored to incorporate any recurrent black-box model or interpretable decision model. We assess CPR through studies on simulated and real data, achieving state-of-the-art performance on the canonical tasks of predicting antibiotic prescription in intensive care units (\$+22{\textbackslash}\%\$ AUROC vs. previous SOTA) and predicting MRI prescription for Alzheimer's patients (\$+7.7{\textbackslash}\%\$ AUROC vs. previous SOTA). With this improvement in predictive performance, CPR closes the accuracy gap between interpretable and black-box methods for policy learning, allowing high-resolution exploration and analysis of context-specific decision models.},
-	urldate = {2023-11-02},
-	publisher = {arXiv},
-	author = {Deuschel, Jannik and Ellington, Caleb N. and Lengerich, Benjamin J. and Luo, Yingtao and Friederich, Pascal and Xing, Eric P.},
-	month = oct,
-	year = {2023},
-	note = {arXiv:2310.07918 [cs, stat]},
-	keywords = {Computer Science - Artificial Intelligence, Computer Science - Machine Learning, Statistics - Machine Learning},
-	file = {arXiv Fulltext PDF:/Users/calebellington/Zotero/storage/AJ8XNWW6/Deuschel et al. - 2023 - Contextualized Policy Recovery Modeling and Inter.pdf:application/pdf},
-}
-
-@misc{lengerich_contextualized_2023,
-	title = {Contextualized {Machine} {Learning}},
-	url = {http://arxiv.org/abs/2310.11340},
-	doi = {10.48550/arXiv.2310.11340},
[…]
