+++ joss/paper.bib
@@ -27,7 +27,7 @@ benefit most from treatments against thrombosis. This approach provides an
 automated methodology to discover heterogeneous and individualized
 effectiveness of treatments.},
 	journal = {J. Biomed. Inform.},
-	author = {Lengerich, Benjamin J and Nunnally, Mark E and Aphinyanaphongs, Yin and Ellington, Caleb and Caruana, Rich},
+	author = {Lengerich, Benjamin J. and Nunnally, Mark E. and Aphinyanaphongs, Yin and Ellington, Caleb N. and Caruana, Rich},
 	month = apr,
 	year = {2022},
 	keywords = {COVID-19, Heterogeneous Treatment Effects, Interpretable Machine Learning, notion, Personalized Medicine},
@@ -91,7 +91,7 @@ Section: New Results},
 	abstract = {Context-specific Bayesian networks (i.e. directed acyclic graphs, DAGs) identify context-dependent relationships between variables, but the non-convexity induced by the acyclicity requirement makes it difficult to share information between context-specific estimators (e.g. with graph generator functions). For this reason, existing methods for inferring context-specific Bayesian networks have favored breaking datasets into subsamples, limiting statistical power and resolution, and preventing the use of multidimensional and latent contexts. To overcome this challenge, we propose NOTEARS-optimized Mixtures of Archetypal DAGs (NOTMAD). NOTMAD models context-specific Bayesian networks as the output of a function which learns to mix archetypal networks according to sample context. The archetypal networks are estimated jointly with the context-specific networks and do not require any prior knowledge. We encode the acyclicity constraint as a smooth regularization loss which is back-propagated to the mixing function; in this way, NOTMAD shares information between context-specific acyclic graphs, enabling the estimation of Bayesian network structures and parameters at even single-sample resolution. We demonstrate the utility of NOTMAD and sample-specific network inference through analysis and experiments, including patient-specific gene expression networks which correspond to morphological variation in cancer.},
 	urldate = {2024-03-14},
 	publisher = {arXiv},
-	author = {Lengerich, Ben and Ellington, Caleb and Aragam, Bryon and Xing, Eric P. and Kellis, Manolis},
+	author = {Lengerich, Benjamin J. and Ellington, Caleb N. and Aragam, Bryon and Xing, Eric P. and Kellis, Manolis},
 	month = nov,
 	year = {2021},
 	note = {arXiv:2111.01104 [cs, stat]},
