+++ joss/paper.bib
@@ -27,7 +27,7 @@ benefit most from treatments against thrombosis. This approach provides an
 automated methodology to discover heterogeneous and individualized
 effectiveness of treatments.},
 	journal = {J. Biomed. Inform.},
-	author = {Lengerich, Benjamin J. and Nunnally, Mark E. and Aphinyanaphongs, Yin and Ellington, Caleb N. and Caruana, Rich},
+	author = {Lengerich, Benjamin J. and Nunnally, Mark E. and Aphinyanaphongs, Yin and Ellington, Caleb and Caruana, Rich},
 	month = apr,
 	year = {2022},
 	keywords = {COVID-19, Heterogeneous Treatment Effects, Interpretable Machine Learning, notion, Personalized Medicine},
@@ -59,7 +59,7 @@ effectiveness of treatments.},
 	abstract = {We examine Contextualized Machine Learning (ML), a paradigm for learning heterogeneous and context-dependent effects. Contextualized ML estimates heterogeneous functions by applying deep learning to the meta-relationship between contextual information and context-specific parametric models. This is a form of varying-coefficient modeling that unifies existing frameworks including cluster analysis and cohort modeling by introducing two reusable concepts: a context encoder which translates sample context into model parameters, and sample-specific model which operates on sample predictors. We review the process of developing contextualized models, nonparametric inference from contextualized models, and identifiability conditions of contextualized models. Finally, we present the open-source PyTorch package ContextualizedML.},
 	urldate = {2023-11-02},
 	publisher = {arXiv},
-	author = {Lengerich, Benjamin and Ellington, Caleb N. and Rubbi, Andrea and Kellis, Manolis and Xing, Eric P.},
+	author = {Lengerich, Benjamin J. and Ellington, Caleb N. and Rubbi, Andrea and Kellis, Manolis and Xing, Eric P.},
 	month = oct,
 	year = {2023},
 	note = {arXiv:2310.11340 [cs, stat]},
