### 2024-03-16 #6469 @cnellington
I authored the majority of our core Python packages under contextualized/, but out codebase also contains many .ipynb files for our documentation, website, and tutorials built with jupyter-book. @blengerich, in addition to making many contributions to our code codebase, also spearheaded our website/documentation efforts, and also did a large amount of standardized style formatting with `black` on our codebase. The line count differences are mainly due to the differences in .ipynb and .py file types, a few reorganizations of large notebook files into new directories, and automated formatting of large swathes of the codebase.

Notebook inflation examples:
- https://github.com/cnellington/Contextualized/pull/121/files
- https://github.com/cnellington/Contextualized/pull/162/files
Style formatting inflation:
- https://github.com/cnellington/Contextualized/pull/126/files

Eric P. Xing is my PhD advisor, and Manolis Kellis is @blengerich's postdoc advisor. While they have not made direct contributions to the code, they have been essential in supporting and directing the software development.
