### 2024-03-16 cnellington/Contextualized#230 @holl-
Easy demo
I'm trying to understand your [easy demo](https://github.com/cnellington/Contextualized/blob/joss/docs/models/easy_regression.ipynb) but am having some trouble. Could you please add more documentation to explain what the code is doing? For example, clearly state your problem in the beginning. What variables are we actually interested in?

Could you also highlight why using your library is easier than doing this with Scikit learn or PyTorch or similar libraries? Couldn't I just pass (C,X) to a regular model and train it in supervised fashion? What does your library do that I couldn't easily achieve otherwise?

You introduce ε in the equation but I don't see this implemented in the code. The other variables are not explained either.
