### 2024-03-16 #6469 @holl-
@cnellington I have some comments for your easy demo notebook, see [this issue](https://github.com/cnellington/Contextualized/issues/230)
