### 2024-03-17 cnellington/Contextualized#230 @cnellington
Hi @holl- thanks for the comment. The goal of Contextualized is to recover context-dependent models from data. The "[Death to Cluster Models](https://contextualized.ml/docs/demos/custom_models.html)" demo digs a little deeper than the demo you mention, and directly compares Contextualized regression against SKLearn regression and simple PyTorch neural networks applied to the same problem.

In general though, traditional statistical models do not account for context-dependent parameters and cannot generalize to new contexts. Neural networks are also insufficient on their own, and do not explicitly reveal a context-specific data distribution. Contextualized models recover context-specific models and generalize to new contexts.

Thanks for the catch about $\epsilon$, we should define this. I will update the demo soon to clarify. $\epsilon$ is a latent centered noise variable in linear regression and can be ignored when predicting the expected value of Y given X. We take our notation from https://en.wikipedia.org/wiki/Linear_regression
