### 2024-03-18 cnellington/Contextualized#230 @holl-
@cnellington Right, the other notebook is more detailed. However, the way your documentation is set up, most people will first look at the easy demo. It would be very helpful if there were some more comments explaining what is going on and linking to further documentation.
