### 2024-03-18 #6469 @fabian-s
> @cnellington Your paper is missing an overview of the _State of the field_. In the 'Statement of need', could you add more concrete motivation for using this library and relate it to existing software solutions (incl. appropriate references)?

Would like to strongly reinforce this and point out this missing reference from the stats literature / R package that seems highly relevant in this context --

`partykit`
https://www.jmlr.org/papers/v16/hothorn15a.html
implements, inter alia, trees and RFs with regression/classification models in each terminal node
