### 2024-03-20 #6469 @cnellington
@fabian-s @holl- Thanks for the clarification, we previously had a more detailed and formal comparison of related works but cut this after seeing the example "Gaia" submission provides only a high-level overview of the field by citing current usage. Most related work to Contextualized can be described by the generic method categories mentioned in our Statement of Need Section 1 (e.g. `partykit` and regression trees are a specific type of mixture modeling). I will re-add our formal problem definition and related works sections to clarify. Right now I am traveling, but you can expect an updated manuscript by next week at the latest.
