+++ joss/paper.bib
@@ -180,3 +180,115 @@ Publication Title: arXiv [cs.LG]},
 	pages = {3000--3008},
 	file = {Full Text PDF:/Users/calebellington/Zotero/storage/9L3UCEBI/Stoica et al. - 2020 - Contextual Parameter Generation for Knowledge Grap.pdf:application/pdf},
 }
+
+@article{fan_statistical_1999,
+	title = {Statistical estimation in varying coefficient models},
+	volume = {27},
+	issn = {0090-5364, 2168-8966},
+	url = {https://projecteuclid.org/journals/annals-of-statistics/volume-27/issue-5/Statistical-estimation-in-varying-coefficient-models/10.1214/aos/1017939139.full},
+	doi = {10.1214/aos/1017939139},
+	abstract = {Varying coefficient models are a useful extension of classical linear models. They arise naturally when one wishes to examine how regression coefficients change over different groups characterized by certain covariates such as age. The appeal of these models is that the coef .cient functions can easily be estimated via a simple local regression.This yields a simple one-step estimation procedure. We show that such a one-step method cannot be optimal when different coefficient functions admit different degrees of smoothness. This drawback can be repaired by using our proposed two-step estimation procedure.The asymptotic mean-squared error for the two-step procedure is obtained and is shown to achieve the optimal rate of convergence. A few simulation studies show that the gain by the two-step procedure can be quite substantial.The methodology is illustrated by an application to an environmental data set.},
+	number = {5},
+	urldate = {2022-06-27},
+	journal = {The Annals of Statistics},
+	author = {Fan, Jianqing and Zhang, Wenyang},
+	month = oct,
+	year = {1999},
+	note = {Publisher: Institute of Mathematical Statistics},
+	keywords = {62G07, 62J12, local linear fit, mean-squared errors, Optimal rate of convergence, personalization, statistics, varying coefficient models},
+	pages = {1491--1518},
+}
+
+@article{kuijjer_estimating_2019,
+	title = {Estimating {Sample}-{Specific} {Regulatory} {Networks}},
+	volume = {14},
+	issn = {2589-0042},
+	doi = {10.1016/j.isci.2019.03.021},
+	abstract = {Biological systems are driven by intricate interactions among molecules. Many methods have been developed that draw on large numbers of expression samples to infer connections between genes (or their products). The result is an aggregate network representing a single estimate for the likelihood of each interaction, or "edge," in the network. Although informative, aggregate models fail to capture population heterogeneity. Here we propose a method to reverse engineer sample-specific networks from aggregate networks. We demonstrate our approach in several contexts, including simulated, yeast microarray, and human lymphoblastoid cell line RNA sequencing data. We use these sample-specific networks to study changes in network topology across time and to characterize shifts in gene regulation that were not apparent in the expression data. We believe that generating sample-specific networks will greatly facilitate the application of network methods to large, complex, and heterogeneous multi-omic datasets, supporting the emerging field of precision network medicine.},
+	language = {eng},
+	journal = {iScience},
+	author = {Kuijjer, Marieke Lydia and Tung, Matthew George and Yuan, GuoCheng and Quackenbush, John and Glass, Kimberly},
+	month = apr,
+	year = {2019},
+	pmid = {30981959},
+	pmcid = {PMC6463816},
+	keywords = {Bioinformatics, Biological Sciences, Complex Systems},
+	pages = {226--240},
+	file = {Full Text:/Users/calebellington/Zotero/storage/MLXVELW9/Kuijjer et al. - 2019 - Estimating Sample-Specific Regulatory Networks.pdf:application/pdf},
+}
+
+@article{kolar_estimating_2010,
+	title = {Estimating time-varying networks},
+	volume = {4},
+	issn = {1932-6157},
+	url = {http://arxiv.org/abs/0812.5087},
[…]
