### 2024-03-26 31bf3f7
typos, equations
