### 2024-03-26 1b76a0a
bib typos, equation newline
