+++ joss/paper.bib
@@ -1,4 +1,3 @@
-
 @article{lengerich_automated_2022,
 	title = {Automated {Interpretable} {Discovery} of {Heterogeneous} {Treatment} {Effectiveness}: {A} {COVID}-19 {Case} {Study}},
 	issn = {1532-0464},
