### 2024-03-26 d2cfa02
fix theta latex codes
