+++ joss/paper.bib
@@ -293,3 +293,34 @@ http://www.sailing.cs.cmu.edu/. CONTACT: epxing@cs.cmu.edu.},
 	year = {2011},
 	pages = {i196--204},
 }
+
+@article{hothorn_partykit_2015,
+	title = {partykit: {A} {Modular} {Toolkit} for {Recursive} {Partytioning} in {R}},
+	volume = {16},
+	issn = {1533-7928},
+	shorttitle = {partykit},
+	url = {http://jmlr.org/papers/v16/hothorn15a.html},
+	abstract = {The R package partykit provides a flexible toolkit for learning, representing, summarizing, and visualizing a wide range of tree- structured regression and classification models. The functionality encompasses: (a) basic infrastructure for representing trees (inferred by any algorithm) so that unified print/plot/predict methods are available; (b) dedicated methods for trees with constant fits in the leaves (or terminal nodes) along with suitable coercion functions to create such trees (e.g., by rpart, RWeka, PMML); (c) a reimplementation of conditional inference trees (ctree, originally provided in the party package); (d) an extended reimplementation of model-based recursive partitioning (mob, also originally in party) along with dedicated methods for trees with parametric models in the leaves. Here, a brief overview of the package and its design is given while more detailed discussions of items (a)—(d) are available in vignettes accompanying the package.},
+	number = {118},
+	urldate = {2024-03-23},
+	journal = {Journal of Machine Learning Research},
+	author = {Hothorn, Torsten and Zeileis, Achim},
+	year = {2015},
+	keywords = {notion},
+	pages = {3905--3909},
+	file = {Full Text PDF:/Users/calebellington/Zotero/storage/S9N4P6K3/Hothorn and Zeileis - 2015 - partykit A Modular Toolkit for Recursive Partytio.pdf:application/pdf},
+}
+
+
+@inproceedings{kolar_sparsistent_2009,
+	title = {Sparsistent {Learning} of {Varying}-coefficient {Models} with {Structural} {Changes}},
+	volume = {22},
+	url = {https://proceedings.neurips.cc/paper/2009/hash/7bcdf75ad237b8e02e301f4091fb6bc8-Abstract.html},
+	abstract = {To estimate the changing structure of a varying-coefficient   varying-structure (VCVS) model remains an important and open problem   in dynamic system modelling, which includes learning trajectories of   stock prices, or uncovering the topology of an evolving gene   network. In this paper, we investigate sparsistent learning of a   sub-family of this model --- piecewise constant VCVS models. We   analyze two main issues in this problem: inferring time points where   structural changes occur and estimating model structure (i.e., model   selection) on each of the constant segments. We propose a two-stage   adaptive procedure, which first identifies jump points of structural   changes and then identifies relevant covariates to a response on   each of the segments. We provide an asymptotic analysis of the   procedure, showing that with the increasing sample size, number of   structural changes, and number of variables, the true model can be   consistently selected. We demonstrate the performance of the method   on synthetic data and apply it to the brain computer interface   dataset. We also consider how this applies to structure estimation   of time-varying probabilistic graphical models.},
+	urldate = {2024-03-26},
+	booktitle = {Advances in {Neural} {Information} {Processing} {Systems}},
+	publisher = {Curran Associates, Inc.},
+	author = {Kolar, Mladen and Song, Le and Xing, Eric},
+	year = {2009},
+	file = {Full Text PDF:/Users/calebellington/Zotero/storage/NUTZZVA9/Kolar et al. - 2009 - Sparsistent Learning of Varying-coefficient Models.pdf:application/pdf},
+}
