+++ joss/paper.bib
@@ -218,26 +218,6 @@ Pages: 2020.06.25.20140053},
 	file = {Full Text:/Users/calebellington/Zotero/storage/MLXVELW9/Kuijjer et al. - 2019 - Estimating Sample-Specific Regulatory Networks.pdf:application/pdf},
 }
 
-@article{kolar_estimating_2010,
-	title = {Estimating time-varying networks},
-	volume = {4},
-	issn = {1932-6157},
-	url = {http://arxiv.org/abs/0812.5087},
-	doi = {10.1214/09-AOAS308},
-	abstract = {Stochastic networks are a plausible representation of the relational information among entities in dynamic systems such as living cells or social communities. While there is a rich literature in estimating a static or temporally invariant network from observation data, little has been done toward estimating time-varying networks from time series of entity attributes. In this paper we present two new machine learning methods for estimating time-varying networks, which both build on a temporally smoothed \$l\_1\$-regularized logistic regression formalism that can be cast as a standard convex-optimization problem and solved efficiently using generic solvers scalable to large networks. We report promising results on recovering simulated time-varying networks. For real data sets, we reverse engineer the latent sequence of temporally rewiring political networks between Senators from the US Senate voting records and the latent evolving regulatory networks underlying 588 genes across the life cycle of Drosophila melanogaster from the microarray time course.},
-	number = {1},
-	urldate = {2024-03-26},
-	journal = {The Annals of Applied Statistics},
-	author = {Kolar, Mladen and Song, Le and Ahmed, Amr and Xing, Eric P.},
-	month = mar,
-	year = {2010},
-	note = {arXiv:0812.5087 [q-bio, stat]},
-	keywords = {Quantitative Biology - Molecular Networks, Quantitative Biology - Quantitative Methods, Statistics - Applications, Statistics - Machine Learning, Statistics - Methodology},
-	annote = {Comment: Published in at http://dx.doi.org/10.1214/09-AOAS308 the Annals of Applied Statistics (http://www.imstat.org/aoas/) by the Institute of Mathematical Statistics (http://www.imstat.org)},
-	file = {arXiv Fulltext PDF:/Users/calebellington/Zotero/storage/9PS5EVN5/Kolar et al. - 2010 - Estimating time-varying networks.pdf:application/pdf},
-}
-
-
 @article{wang_bayesian_2022,
 	title = {Bayesian {Edge} {Regression} in {Undirected} {Graphical} {Models} to {Characterize} {Interpatient} {Heterogeneity} in {Cancer}},
 	volume = {117},
@@ -255,7 +235,6 @@ Pages: 2020.06.25.20140053},
 	pages = {533--546},
 	file = {Accepted Version:/Users/calebellington/Zotero/storage/A7P25CK7/Wang et al. - 2022 - Bayesian Edge Regression in Undirected Graphical M.pdf:application/pdf},
 }
-
 
 @article{parikh_treegl_2011,
 	title = {{TREEGL}: reverse engineering tree-evolving gene networks underlying developing biological lineages},
@@ -294,23 +273,6 @@ http://www.sailing.cs.cmu.edu/. CONTACT: epxing@cs.cmu.edu.},
 	pages = {i196--204},
 }
 
-@article{hothorn_partykit_2015,
-	title = {partykit: {A} {Modular} {Toolkit} for {Recursive} {Partytioning} in {R}},
-	volume = {16},
-	issn = {1533-7928},
-	shorttitle = {partykit},
-	url = {http://jmlr.org/papers/v16/hothorn15a.html},
[…]
