### 2024-03-26 64d7dfa
update tv networks with appropriate reference
