+++ joss/paper.bib
@@ -273,20 +273,24 @@ http://www.sailing.cs.cmu.edu/. CONTACT: epxing@cs.cmu.edu.},
 	pages = {i196--204},
 }
 
-
-@inproceedings{kolar_sparsistent_2009,
-	title = {Sparsistent {Learning} of {Varying}-coefficient {Models} with {Structural} {Changes}},
-	volume = {22},
-	url = {https://proceedings.neurips.cc/paper/2009/hash/7bcdf75ad237b8e02e301f4091fb6bc8-Abstract.html},
-	abstract = {To estimate the changing structure of a varying-coefficient   varying-structure (VCVS) model remains an important and open problem   in dynamic system modelling, which includes learning trajectories of   stock prices, or uncovering the topology of an evolving gene   network. In this paper, we investigate sparsistent learning of a   sub-family of this model --- piecewise constant VCVS models. We   analyze two main issues in this problem: inferring time points where   structural changes occur and estimating model structure (i.e., model   selection) on each of the constant segments. We propose a two-stage   adaptive procedure, which first identifies jump points of structural   changes and then identifies relevant covariates to a response on   each of the segments. We provide an asymptotic analysis of the   procedure, showing that with the increasing sample size, number of   structural changes, and number of variables, the true model can be   consistently selected. We demonstrate the performance of the method   on synthetic data and apply it to the brain computer interface   dataset. We also consider how this applies to structure estimation   of time-varying probabilistic graphical models.},
+@article{kolar_estimating_2010,
+	title = {Estimating time-varying networks},
+	volume = {4},
+	issn = {1932-6157},
+	url = {http://arxiv.org/abs/0812.5087},
+	doi = {10.1214/09-AOAS308},
+	abstract = {Stochastic networks are a plausible representation of the relational information among entities in dynamic systems such as living cells or social communities. While there is a rich literature in estimating a static or temporally invariant network from observation data, little has been done toward estimating time-varying networks from time series of entity attributes. In this paper we present two new machine learning methods for estimating time-varying networks, which both build on a temporally smoothed \$l\_1\$-regularized logistic regression formalism that can be cast as a standard convex-optimization problem and solved efficiently using generic solvers scalable to large networks. We report promising results on recovering simulated time-varying networks. For real data sets, we reverse engineer the latent sequence of temporally rewiring political networks between Senators from the US Senate voting records and the latent evolving regulatory networks underlying 588 genes across the life cycle of Drosophila melanogaster from the microarray time course.},
+	number = {1},
 	urldate = {2024-03-26},
-	booktitle = {Advances in {Neural} {Information} {Processing} {Systems}},
-	publisher = {Curran Associates, Inc.},
-	author = {Kolar, Mladen and Song, Le and Xing, Eric},
-	year = {2009},
-	file = {Full Text PDF:/Users/calebellington/Zotero/storage/NUTZZVA9/Kolar et al. - 2009 - Sparsistent Learning of Varying-coefficient Models.pdf:application/pdf},
+	journal = {The Annals of Applied Statistics},
+	author = {Kolar, Mladen and Song, Le and Ahmed, Amr and Xing, Eric P.},
+	month = mar,
+	year = {2010},
+	note = {arXiv:0812.5087 [q-bio, stat]},
+	keywords = {Quantitative Biology - Molecular Networks, Quantitative Biology - Quantitative Methods, Statistics - Applications, Statistics - Machine Learning, Statistics - Methodology},
[…]
