+++ joss/paper.md
@@ -69,11 +69,15 @@ With the flexibility of context-dependent parameters, each context-specific mode
 In its basic form: $x_i \sim P(X_i ;θ_i)$, where $i$ indexes a sample, $θ_i$ is the parameters defining the sample-specific distribution, and $x_i$ corresponds to the observation drawn from this sample-specific distribution, where understanding sample heterogeneity is equivalent to estimating data distributions with sample-specific parameters.
 Some methods, such as sample-left-out models [@kuijjer_estimating_2019], provide sample-specific estimators without additional information but lack desirable statistical properties such as the ability to generalize to new samples or test model performance on held-out data.
 Due to the difficulty of estimating sample-specific parameters, most methods make use of side information, covariates or “context,” as an indicator of sample-to-sample variation [@hastie_varying-coefficient_1993; @fan_statistical_1999; @wang_bayesian_2022; @kolar_estimating_2010; @parikh_treegl_2011]. 
-However, prior methods only permit model variation across a few continuous covariates [@hastie_varying-coefficient_1993; @fan_statistical_1999; @wang_bayesian_2022], or a small number of groups [@kolar_estimating_2010; @parikh_treegl_2011; @zeileis_model-based_2008], and do not scale to high-dimensional, continuous, and sample-specific variation.
-Recently, contextual explanation networks (CENs) reframed the sample-specific parameter estimation problem as a more flexible and generalizable latent variable inference problem that can naturally leverage any additional contextual data pertinent to the study subjects [@al-shedivat_contextual_2020]. 
-Contextualized Machine Learning is a generalization of the CEN method, which provides a unified mathematical framework for inferring and estimating personalized models of heterogeneous and context-dependent systems [@lengerich_contextualized_2023].
+However, prior methods only permit limited use of contextual side information, allowing models to vary over a few continuous covariates [@hastie_varying-coefficient_1993; @fan_statistical_1999; @wang_bayesian_2022], or a small number of groups [@kolar_estimating_2010; @parikh_treegl_2011; @zeileis_model-based_2008], and do not scale to high-dimensional, complex, and sample-specific variation.
+`partykit` [@hothorn_partykit_2015] goes beyond this by enabling random forests of model-based recursive partitioning [@zeileis_model-based_2008] on contextual information.
+This method allows learning complex non-linear relationships between contextual information and linear regression parameters, but only permits linear model personalization.
+Additionally, the relationship between linear coefficients and contextual information must be represented using tree-based methods, which struggle with high-dimensional and non-tabular data types.
+Recently, the contextual explanation network (CEN) was developed to learn this context-model relationship using a deep neural network, benefiting from a wide range of architectures targeting high-dimensional non-tabular data [@al-shedivat_contextual_2020].
+However, the CEN, like model-based partitioning, is designed only for linear model personalization.
+Contextualized Machine Learning is the generalization of the CEN method, which reframes the sample-specific parameter estimation problem as a more flexible and generalizable latent variable inference problem which provides a unified mathematical framework for inferring and estimating personalized models of heterogeneous and context-dependent systems using a wide range of model types [@lengerich_contextualized_2023].
 
-Formally, given subject data $X = \{X_i\}_{i=1}^N$ and context data $C = \{C_i\}_{i=1}^N$ where $i$ indexes samples, we can express the likelihood of all data in the form of 
+Formally, Contexutalized Machine Learning uses subject data $X = \{X_i\}_{i=1}^N$ and context data $C = \{C_i\}_{i=1}^N$ where $i$ indexes samples, we can express the likelihood of all data in the form of 
[…]
