### 2024-03-26 #6469 @cnellington
@fabian-s @holl- I have updated the statement of need to include a formal problem definition and an overview of related work. The only missing DOI is for partykit, where the JMLR page did not have a DOI and the ACM DOI was invalid.

We will continue improving test coverage where possible while we wait for your feedback.
