### 2024-03-27 #6469 @holl-
@cnellington The paper looks good now.

Could you explain the reason for having an `easy` module in your API?
This seems to imply that everything else is hard to do. Having an easy start with a library is good but people shouldn't have to learn a different system for more involved tasks.

Assuming you have written an analysis using `easy`, is it possible to obtain the same results without using the `easy` API (possibly with more code)?
