### 2024-03-28 #6469 @cnellington
@holl- The `easy` and `analysis` modules contain our user-facing APIs. The other modules are developer-facing, mainly written with PyTorch Lightning API patterns. The `easy` module internalizes training/evaluation of these PyTorch modules and exposes this through an SKLearn-like API wrapper. By definition, anything done in `easy` or `analysis` can be done with the developer-facing modules, but we do not expect users to know or use the other modules. Among developers, we do make a point to keep our private codebase in line with the PyTorch ecosystem so `contextualized` benefits internally from tools produced there.
