### 2024-03-28 #6469 @holl-
@cnellington Okay, then how about calling that module `model` / `models` instead of `easy`?

Also I noticed that some classes have a plural name, e.g. `ContextualizedBayesianNetworks`. This is confusing. Does instantiating this class create multiple networks? If so, how about calling it a forest instead?

The documentation does not really explain the parameters, e.g. what are the bootstraps and archetypes? Could you add explanations or link to a page that introduces these terms?
