### 2024-04-05 #6469 @pescap
> @holl- The `easy` and `analysis` modules contain our user-facing APIs. The other modules are developer-facing, mainly written with PyTorch Lightning API patterns. The `easy` module internalizes training/evaluation of these PyTorch modules and exposes this through an SKLearn-like API wrapper. By definition, anything done in `easy` or `analysis` can be done with the developer-facing modules, but we do not expect users to know or use the other modules. Among developers, we do make a point to keep our private codebase in line with the PyTorch ecosystem so `contextualized` benefits internally from tools produced there.

I had the same question.

Then I read in [easy_regression.ipynb](https://github.com/cnellington/Contextualized/blob/main/docs/models/easy_regression.ipynb):

> Note: These contextualized.easy interfaces don't try to use best PyTorch practices (prioritizing easy use over computational efficiency), so it's recommended to use contextualized models for any heavy lifting (see this [demo](https://github.com/cnellington/Contextualized/blob/main/docs/demos/custom_models.ipynb)).

It would be beneficial to add this note (or a similar explanation) before in the [docs](https://contextualized.ml/docs/intro.html) or README.md.

What do you think about that?
