### 2024-04-07 49aff48
add network quicktests to main test file, fix bug with averaged correlation in ContextualizedCorrelationNetworks
