+++ tests.py
@@ -3,6 +3,9 @@ from contextualized.regression.tests import *
 from contextualized.dags.tests import *
 from contextualized.easy.tests.test_regressor import *
 from contextualized.easy.tests.test_classifier import *
+from contextualized.easy.tests.test_markov_networks import *
+from contextualized.easy.tests.test_correlation_networks import *
+from contextualized.easy.tests.test_bayesian_networks import *
 
 if __name__ == '__main__':
     unittest.main()
