### 2024-04-07 529672e
remove rho < 0 assertion, not guaranteed
