### 2024-04-09 04bcf33
update analysis tests with linear encoder, add to all tests
