+++ tests.py
@@ -6,6 +6,7 @@ from contextualized.easy.tests.test_classifier import *
 from contextualized.easy.tests.test_markov_networks import *
 from contextualized.easy.tests.test_correlation_networks import *
 from contextualized.easy.tests.test_bayesian_networks import *
+from contextualized.analysis.tests import *
 
 if __name__ == '__main__':
     unittest.main()
