### 2024-04-17 f84c12a
remove communication time between workers in dag tests, add dag tests_fast.py
