### 2024-04-19 b471c12
added select_good_bootstraps tests and reference, fixed model list indexing bug
