### 2024-04-19 6d93fe9
add gam tests, move model fitting out of test setUp (model fitting and multithreading don't agree)
