+++ contextualized/analysis/tests.py
@@ -4,6 +4,7 @@ Unit tests for analysis utilities.
 
 import unittest
 import copy
+import torch
 import numpy as np
 import pandas as pd
 
@@ -25,40 +26,33 @@ class TestTestEachContext(unittest.TestCase):
 		"""
 		Shared data setup.
 		"""
+		torch.manual_seed(0)
+		np.random.seed(0)
 		n_samples = 1000
 		C = np.random.uniform(0, 1, size=(n_samples, 2))
 		X = np.random.uniform(0, 1, size=(n_samples, 2))
 		beta = np.concatenate([np.ones((n_samples, 1)), C], axis=1)
-		Y = np.sum(beta[:, :2] * X, axis=1)
+		Y = np.sum(beta[:, :2] * X, axis=1)  # X1 changes effect under C0. C1 has no effect, X0 is constant
 
 		self.C_train_df = pd.DataFrame(C, columns=['C0', 'C1'])
 		self.X_train_df = pd.DataFrame(X, columns=['X0', 'X1'])
 		self.Y_train_df = pd.DataFrame(Y, columns=['Y'])
 
-		self.pvals = test_each_context(ContextualizedRegressor, self.C_train_df, self.X_train_df, self.Y_train_df, model_kwargs={'encoder_type': 'mlp', 'layers': 0}, fit_kwargs={'max_epochs': 1, 'learning_rate': 1e-2, 'n_bootstraps': 40})
-
-	def test_output_shape(self):
+	def test_test_each_context(self):
 		"""
 		Test that the output shape of the test_each_context function is as expected.
 		"""
+		pvals = test_each_context(ContextualizedRegressor, self.C_train_df, self.X_train_df, self.Y_train_df, model_kwargs={'encoder_type': 'mlp', 'layers': 1}, fit_kwargs={'max_epochs': 1, 'learning_rate': 1e-2, 'n_bootstraps': 10})
+
 		expected_shape = (self.C_train_df.shape[1] * self.X_train_df.shape[1] * self.Y_train_df.shape[1], 4)
-		self.assertEqual(self.pvals.shape, expected_shape)
-	
-	def test_valid_pval_range(self):
-		"""
-		Test that all pvals are in the valid range.
-		"""
-		self.assertTrue(all(0 <= pval <= 1 for pval in self.pvals['Pvals']))
+		self.assertEqual(pvals.shape, expected_shape)	
+		self.assertTrue(all(0 <= pval <= 1 for pval in pvals['Pvals']))
 
-	def test_expected_significant_pval(self):
-		"""
-		Test that expected significant pvals are in fact significant.
-		"""
-		pval_c0_x1 = self.pvals.loc[1, 'Pvals']
-		self.assertTrue(pval_c0_x1 < 0.05, "C0 X1 p-value is not significant.")
+		pval_c0_x1 = pvals.loc[1, 'Pvals']
+		self.assertTrue(pval_c0_x1 < 0.2, "C0 X1 p-value is not significant.")
 
-		other_pvals = self.pvals.drop(1)
-		self.assertTrue(all(pval >= 0.05 for pval in other_pvals['Pvals']), "Other p-values are significant.")
+		other_pvals = pvals.drop(1)
+		self.assertTrue(all(pval >= 0.2 for pval in other_pvals['Pvals']), "Other p-values are significant.")
 
 
 class TestSelectGoodBootstraps(unittest.TestCase):
@@ -67,26 +61,26 @@ class TestSelectGoodBootstraps(unittest.TestCase):
 		super().__init__(*args, **kwargs)
 	
 	def setUp(self):
-		self.model = ContextualizedRegressor(n_bootstraps = 3)
 		C = np.random.uniform(0, 1, size=(100, 2))
[…]
