+++ contextualized/easy/tests/test_classifier.py
@@ -5,10 +5,10 @@ Unit tests for Easy Classifier.
 import unittest
 import numpy as np
 
-from contextualized.easy import ContextualizedClassifier
+from contextualized.easy import ContextualizedClassifier, ContextualGAMClassifier
 
 
-class TestEasyClassifier(unittest.TestCase):
+class TestEasyClassifiers(unittest.TestCase):
     """
     Test Easy Classifier models.
     """
@@ -44,6 +44,20 @@ class TestEasyClassifier(unittest.TestCase):
         Y = np.random.binomial(1, 0.5, size=(n_samples, y_dim))
 
         model = ContextualizedClassifier(alpha=1e-1, encoder_type="mlp")
+        self._quicktest(model, C, X, Y, max_epochs=10, es_patience=float("inf"))
+
+    def test_gam_classifier(self):
+        """Test Case for ContextualGAMClassifier."""
+
+        n_samples = 1000
+        c_dim = 100
+        x_dim = 3
+        y_dim = 1
+        C = np.random.uniform(-1, 1, size=(n_samples, c_dim))
+        X = np.random.uniform(-1, 1, size=(n_samples, x_dim))
+        Y = np.random.binomial(1, 0.5, size=(n_samples, y_dim))
+
+        model = ContextualGAMClassifier(alpha=1e-1, encoder_type="mlp")
         self._quicktest(model, C, X, Y, max_epochs=10, es_patience=float("inf"))
 
 
