### 2024-04-19 faade5c
clean up old bugs moving model out of setUp
