### 2024-04-19 2c98be6
reorganize easy tests, remove depreciated auto_lr_find from baseline dag trainer
