+++ contextualized/easy/tests.py
@@ -1,0 +1,364 @@
+""""
+Unit tests for Easy Networks.
+"""
+
+import unittest
+import numpy as np
+import torch
+
+from contextualized.easy import (
+    ContextualizedMarkovNetworks, 
+    ContextualizedCorrelationNetworks, 
+    ContextualizedBayesianNetworks,
+    ContextualizedRegressor,
+    ContextualizedClassifier,
+    ContextualGAMClassifier,
+    ContextualGAMRegressor,
+)
+from contextualized.utils import DummyParamPredictor, DummyYPredictor
+
+
+class TestEasyNetworks(unittest.TestCase):
+    """
+    Test Easy Network models.
+    """
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def _quicktest(self, model, C, X, **kwargs):
+        print(f"{type(model)} quicktest")
+        model.fit(C, X, max_epochs=0)
+        err_init = model.measure_mses(C, X)
+        model.fit(C, X, **kwargs)
+        err_trained = model.measure_mses(C, X)
+        W_pred = model.predict_networks(C)
+        assert W_pred.shape == (C.shape[0], X.shape[1], X.shape[1])
+        assert np.mean(err_trained) < np.mean(err_init)
+
+
+class TestContextualizedMarkovNetworks(TestEasyNetworks):
+    """
+    Test Contextualized Markov Network models.
+    """
+    def setUp(self):
+        """
+        Shared unit test setup code.
+        """
+        self.n_samples = 100
+        self.c_dim = 4
+        self.x_dim = 5
+        C = torch.rand((self.n_samples, self.c_dim)) - 0.5
+        # TODO: Use graph utils to generate X from a network.
+        X = torch.rand((self.n_samples, self.x_dim)) - 0.5
+        self.C, self.X = C.numpy(), X.numpy()
+
+    def test_markov(self):
+        """Test Case for ContextualizedMarkovNetworks."""
+        model = ContextualizedMarkovNetworks()
+        self._quicktest(model, self.C, self.X, max_epochs=10, learning_rate=1e-3)
+        self._quicktest(
+            model, self.C, self.X, max_epochs=10, learning_rate=1e-3, val_split=0.5
+        )
+
+        model = ContextualizedMarkovNetworks(num_archetypes=16)
+        self._quicktest(model, self.C, self.X, max_epochs=10, learning_rate=1e-3)
+
+        model = ContextualizedMarkovNetworks(encoder_type="ngam", num_archetypes=16)
+        self._quicktest(model, self.C, self.X, max_epochs=10, learning_rate=1e-3)
+        omegas = model.predict_precisions(self.C, individual_preds=False)
+        assert np.shape(omegas) == (self.n_samples, self.x_dim, self.x_dim)
+        omegas = model.predict_precisions(self.C, individual_preds=True)
+        assert np.shape(omegas) == (1, self.n_samples, self.x_dim, self.x_dim)
+
+
+class TestContextualizedCorrelationNetworks(TestEasyNetworks):
+    """
+    Test Contextualized Correlation Network models.
+    """
+    def setUp(self):
+        """
+        Shared unit test setup code.
+        """
+        self.n_samples = 100
+        self.c_dim = 4
+        self.x_dim = 5
+        C = torch.rand((self.n_samples, self.c_dim)) - 0.5
+        # TODO: Use graph utils to generate X from a network.
+        X = torch.rand((self.n_samples, self.x_dim)) - 0.5
+        self.C, self.X = C.numpy(), X.numpy()
+
+    def test_correlation(self):
+        """
[…]
