### 2024-04-19 0bdcbab
remove lambdas from arguments, cannot be pickled, replaced with functools.partial. Added tests to modules and save/load utils
