+++ contextualized/tests.py
@@ -1,0 +1,152 @@
+import os
+import unittest
+import numpy as np
+import torch
+from contextualized.modules import SoftSelect, Explainer, MLP, NGAM, Linear
+from contextualized.easy import ContextualizedRegressor, ContextualizedBayesianNetworks, ContextualizedCorrelationNetworks
+from contextualized.baselines import BayesianNetwork, CorrelationNetwork
+
+
+class TestModules(unittest.TestCase):
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def setUp(self):
+        """
+        Shared data setup.
+        """
+        self.N_SAMPLES = 100
+        self.X_DIM = 10
+        self.Y_DIM = 5
+        self.K_ARCHETYPES = 3
+        self.WIDTH = 50
+        self.LAYERS = 5
+        self.X_data = torch.rand((self.N_SAMPLES, self.X_DIM))
+        self.IN_DIMS = (3, 4)
+        self.OUT_SHAPE = (5, 6)
+        self.Z1 = torch.randn(self.N_SAMPLES, self.IN_DIMS[0])
+        self.Z2 = torch.randn(self.N_SAMPLES, self.IN_DIMS[1])
+
+    def test_mlp(self):
+        """
+        Test that the output shape of the MLP is as expected.
+        """
+        mlp = MLP(self.X_DIM, self.Y_DIM, self.WIDTH, self.LAYERS)
+        assert mlp(self.X_data).shape == (self.N_SAMPLES, self.Y_DIM)
+
+    def test_ngam(self):
+        """
+        Test that the output shape of the NGAM is as expected.
+        """
+        ngam = NGAM(self.X_DIM, self.Y_DIM, self.WIDTH, self.LAYERS)
+        assert ngam(self.X_data).shape == (self.N_SAMPLES, self.Y_DIM)
+
+    def test_softselect(self):
+        """
+        Test that the output shape of the SoftSelect is as expected.
+        """
+        softselect = SoftSelect(self.IN_DIMS, self.OUT_SHAPE)
+        assert softselect(self.Z1, self.Z2).shape == (self.N_SAMPLES, *self.OUT_SHAPE)
+
+        precycle_vals = softselect.archetypes
+        assert precycle_vals.shape == (*self.OUT_SHAPE, *self.IN_DIMS)
+        postcycle_vals = softselect.get_archetypes()
+        assert postcycle_vals.shape == (*self.IN_DIMS, *self.OUT_SHAPE)
+        softselect.set_archetypes(torch.randn(*self.IN_DIMS, *self.OUT_SHAPE))
+        assert (softselect.archetypes != precycle_vals).any()
+        softselect.set_archetypes(postcycle_vals)
+        assert (softselect.archetypes == precycle_vals).all()
+
+    def test_explainer(self):
+        explainer = Explainer(self.IN_DIMS[0], self.OUT_SHAPE)
+        ret = explainer(self.Z1)
+
+        precycle_vals = explainer.archetypes
+        assert precycle_vals.shape == (*self.OUT_SHAPE, self.IN_DIMS[0])
+        postcycle_vals = explainer.get_archetypes()
+        assert postcycle_vals.shape == (self.IN_DIMS[0], *self.OUT_SHAPE)
+        explainer.set_archetypes(torch.randn(self.IN_DIMS[0], *self.OUT_SHAPE))
+        assert (explainer.archetypes != precycle_vals).any()
+        explainer.set_archetypes(postcycle_vals)
+        assert (explainer.archetypes == precycle_vals).all()
+
+    def test_linear(self):
+        linear_encoder = Linear(self.X_DIM, self.Y_DIM)
[…]
