+++ contextualized/modules.py
@@ -155,53 +155,3 @@ class Linear(nn.Module):
 
 
 ENCODERS = {"mlp": MLP, "ngam": NGAM, "linear": Linear}
-
-
-if __name__ == "__main__":
-    N_SAMPLES = 100
-    X_DIM = 10
-    Y_DIM = 5
-    K_ARCHETYPES = 3
-    WIDTH = 50
-    LAYERS = 5
-    X_data = torch.rand((N_SAMPLES, X_DIM))
-
-    mlp = MLP(X_DIM, Y_DIM, WIDTH, LAYERS)
-    mlp(X_data)
-
-    ngam = NGAM(X_DIM, Y_DIM, WIDTH, LAYERS)
-    ngam(X_data)
-
-    IN_DIMS = (3, 4)
-    OUT_SHAPE = (5, 6)
-    Z1 = torch.randn(N_SAMPLES, IN_DIMS[0])
-    Z2 = torch.randn(N_SAMPLES, IN_DIMS[1])
-    softselect = SoftSelect(IN_DIMS, OUT_SHAPE)
-    softselect(Z1, Z2)
-
-    precycle_vals = softselect.archetypes
-    assert precycle_vals.shape == (*OUT_SHAPE, *IN_DIMS)
-    postcycle_vals = softselect.get_archetypes()
-    assert postcycle_vals.shape == (*IN_DIMS, *OUT_SHAPE)
-    softselect.set_archetypes(torch.randn(*IN_DIMS, *OUT_SHAPE))
-    assert (softselect.archetypes != precycle_vals).any()
-    softselect.set_archetypes(postcycle_vals)
-    assert (softselect.archetypes == precycle_vals).all()
-
-    IN_DIMS = (3,)
-    explainer = Explainer(IN_DIMS[0], OUT_SHAPE)
-    explainer(Z1)
-
-    precycle_vals = explainer.archetypes
-    assert precycle_vals.shape == (*OUT_SHAPE, *IN_DIMS)
-    postcycle_vals = explainer.get_archetypes()
-    assert postcycle_vals.shape == (*IN_DIMS, *OUT_SHAPE)
-    explainer.set_archetypes(torch.randn(*IN_DIMS, *OUT_SHAPE))
-    assert (explainer.archetypes != precycle_vals).any()
-    explainer.set_archetypes(postcycle_vals)
-    assert (explainer.archetypes == precycle_vals).all()
-
-    linear_encoder = Linear(X_DIM, Y_DIM)
-    linear_output = linear_encoder(X_data)
-
-    assert linear_output.shape == (N_SAMPLES, Y_DIM)
