+++ tests.py
@@ -1,11 +1,8 @@
 import unittest
+from contextualized.tests import *
 from contextualized.regression.tests import *
 from contextualized.dags.tests import *
-from contextualized.easy.tests.test_regressor import *
-from contextualized.easy.tests.test_classifier import *
-from contextualized.easy.tests.test_markov_networks import *
-from contextualized.easy.tests.test_correlation_networks import *
-from contextualized.easy.tests.test_bayesian_networks import *
+from contextualized.easy.tests import *
 from contextualized.analysis.tests import *
 from contextualized.baselines.tests import *
 

[…]
