+++ contextualized/dags/losses.py
@@ -75,22 +75,3 @@ def linear_sem_loss(x_true, w_pred):
     """
     x_prime = dag_pred(x_true, w_pred)
     return (0.5 / x_true.shape[0]) * torch.square(torch.linalg.norm(x_true - x_prime))
-
-
-def NOTEARS_loss(x_true, w_pred, l1_lambda, alpha, rho):
-    """Computes NOTEARS loss between true x and predicted network, for torch tensors.
-    Works on batches only.
-
-    Args:
-        w_pred (torch.FloatTensor): _description_
-        alpha (float): Alpha DAG loss param
-        rho (float): Rho DAG loss param
-
-    Returns:
-        torch.tensor: NOTEARS loss for data features and predicted network (batches only).
-    """
-    mse_term = linear_sem_loss(x_true, w_pred)
-    l1_term = l1_loss(w_pred, l1_lambda)
-    dag_term = dag_loss_notears(w_pred, alpha, rho)
-    notears = mse_term + l1_term + dag_term
-    return notears
