### 2024-04-20 69fcb66
add seeds for consistent test behavior
