### 2024-04-21 38cc018
remove dill dependency for issue in python 3.8 with dill>=0.3.6, use partials for pickling. Add torch version cap for segfault issue
