+++ contextualized/tests.py
@@ -90,6 +90,9 @@ class TestSaveLoad(unittest.TestCase):
         C = np.random.uniform(0, 1, size=(100, 2))
         X = np.random.uniform(0, 1, size=(100, 2))
         Y = np.random.uniform(0, 1, size=(100, 2))
+        C2 = np.random.uniform(0, 1, size=(100, 2))
+        X2 = np.random.uniform(0, 1, size=(100, 2))
+        Y2 = np.random.uniform(0, 1, size=(100, 2))
         mlp = MLP(2, 2, 50, 5)
         Y_pred = mlp(torch.Tensor(X)).detach().numpy()
         save(mlp, 'unittest_model.pt')
@@ -97,6 +100,7 @@ class TestSaveLoad(unittest.TestCase):
         mlp_loaded = load('unittest_model.pt')
         Y_pred_loaded = mlp_loaded(torch.Tensor(X)).detach().numpy()
         assert np.all(Y_pred == Y_pred_loaded)
+        os.remove('unittest_model.pt')
 
         model = ContextualizedRegressor()
         model.fit(C, X, Y)
@@ -106,6 +110,15 @@ class TestSaveLoad(unittest.TestCase):
         model_loaded = load('unittest_model.pt')
         Y_pred_loaded = model_loaded.predict(C, X)
         assert np.all(Y_pred == Y_pred_loaded)
+        os.remove('unittest_model.pt')
+        model_loaded.fit(C2, X2, Y2)
+        Y_pred2 = model_loaded.predict(C2, X2)
+        assert not np.all(Y_pred_loaded == Y_pred2)
+        save(model_loaded, 'unittest_model.pt')
+        del model_loaded
+        model_loaded2 = load('unittest_model.pt')
+        Y_pred_loaded2 = model_loaded2.predict(C2, X2)
+        assert np.all(Y_pred2 == Y_pred_loaded2)
         os.remove('unittest_model.pt')
 
         model = ContextualizedBayesianNetworks()
@@ -117,6 +130,15 @@ class TestSaveLoad(unittest.TestCase):
         pred_loaded = model_loaded.predict_networks(C)
         assert np.all(np.array(pred) == np.array(pred_loaded))
         os.remove('unittest_model.pt')
+        model_loaded.fit(C2, X2)
+        pred2 = model_loaded.predict_networks(C2)
+        assert not np.all(np.array(pred_loaded) == np.array(pred2))
+        save(model_loaded, 'unittest_model.pt')
+        del model_loaded
+        model_loaded2 = load('unittest_model.pt')
+        pred_loaded2 = model_loaded2.predict_networks(C2)
+        assert np.all(np.array(pred2) == np.array(pred_loaded2))
+        os.remove('unittest_model.pt')
 
         model = ContextualizedCorrelationNetworks()
         model.fit(C, X)
@@ -126,6 +148,15 @@ class TestSaveLoad(unittest.TestCase):
         model_loaded = load('unittest_model.pt')
         pred_loaded = model_loaded.predict_correlation(C)
         assert np.all(np.array(pred) == np.array(pred_loaded))
+        os.remove('unittest_model.pt')
+        model_loaded.fit(C2, X2)
+        pred2 = model_loaded.predict_correlation(C2)
+        assert not np.all(np.array(pred_loaded) == np.array(pred2))
+        save(model_loaded, 'unittest_model.pt')
+        del model_loaded
+        model_loaded2 = load('unittest_model.pt')
+        pred_loaded2 = model_loaded2.predict_correlation(C2)
+        assert np.all(np.array(pred2) == np.array(pred_loaded2))
         os.remove('unittest_model.pt')
 
         model = BayesianNetwork()
[…]
