### 2024-04-21 010ebfc
add deterministic test seed to easy networks tests
