### 2024-04-21 6bdcc19
format code, run linter, update dev requirements with pylint-badge commit hash
