+++ contextualized/easy/wrappers/SKLearnWrapper.py
@@ -1,6 +1,7 @@
 """
 An sklearn-like wrapper for Contextualized models.
 """
+
 import copy
 import os
 from typing import *
