+++ contextualized/dags/trainers.py
@@ -1,6 +1,7 @@
 """
 PyTorch-Lightning trainers used for Contextualized graphs.
 """
+
 import torch
 import pytorch_lightning as pl
 
