+++ contextualized/regression/tests.py
@@ -1,6 +1,7 @@
 """
 Unit tests for Contextualized Regression.
 """
+
 import unittest
 import numpy as np
 import torch
