### 2024-04-21 #6469 @cnellington
@holl- @pescap @fabian-s  We've expanded our test coverage, including new tests for basic functionality and performance (e.g. [contextualized.analysis.tests.TestPvals](https://github.com/cnellington/Contextualized/blob/12843bbab6bd2f63db1dfdc5247ab865ca47cfa6/contextualized/analysis/tests.py#L106), [contextualized.easy.tests.TestEasyRegression](https://github.com/cnellington/Contextualized/blob/12843bbab6bd2f63db1dfdc5247ab865ca47cfa6/contextualized/easy/tests.py#L260)).

The most recent coverage metrics are below. All the files below 85% coverage can be justified. Embeddings, accuracy_split, and effects are all visualization/reporting utilities and can't be included in unittests, but we do verify their functionality in the [model analysis demo](https://contextualized.ml/docs/model-analysis.html). The only other files below 85% are contextualized.dags.graph_utils, which has several options for noise type that don't need to be tested, and contextualized.functions, where we do not test constant-valued functions (i.e. f(x) = c).

| Module                                       | statements | missing | excluded | coverage |
|----------------------------------------------|------------|---------|----------|----------|
| contextualized/analysis/embeddings.py       | 70         | 62      | 0        | 11%      |
| contextualized/analysis/accuracy_split.py   | 39         | 33      | 0        | 15%      |
| contextualized/analysis/effects.py          | 208        | 143     | 0        | 31%      |
| contextualized/dags/graph_utils.py          | 102        | 23      | 0        | 77%      |
| contextualized/functions.py                 | 32         | 6       | 0        | 81%      |
| contextualized/regression/regularizers.py   | 19         | 2       | 0        | 89%      |
| contextualized/analysis/bootstraps.py       | 10         | 1       | 0        | 90%      |
| contextualized/easy/wrappers/SKLearnWrapper.py | 218    | 17      | 0        | 92%      |
| contextualized/regression/datasets.py       | 86         | 6       | 0        | 93%      |
| contextualized/baselines/networks.py        | 171        | 10      | 0        | 94%      |
| contextualized/easy/ContextualizedRegressor.py | 17     | 1       | 0        | 94%      |
| contextualized/dags/lightning_modules.py    | 152        | 7       | 0        | 95%      |
| contextualized/easy/ContextualizedNetworks.py | 97      | 5       | 0        | 95%      |
| contextualized/baselines/tests.py           | 36         | 1       | 0        | 97%      |
| contextualized/regression/lightning_modules.py | 303   | 10      | 0        | 97%      |
| contextualized/analysis/pvals.py            | 51         | 1       | 0        | 98%      |
| contextualized/analysis/tests.py            | 86         | 1       | 0        | 99%      |
| contextualized/dags/tests.py                | 131        | 1       | 0        | 99%      |
| contextualized/easy/tests.py                | 224        | 1       | 0        | 99%      |
| contextualized/regression/tests.py          | 144        | 1       | 0        | 99%      |
| contextualized/tests.py                     | 165        | 1       | 0        | 99%      |
| contextualized/__init__.py                  | 7          | 0       | 0        | 100%     |
| contextualized/analysis/__init__.py         | 5          | 0       | 0        | 100%     |
| contextualized/baselines/__init__.py        | 1          | 0       | 0        | 100%     |
| contextualized/dags/__init__.py             | 7          | 0       | 0        | 100%     |
| contextualized/dags/losses.py               | 26         | 0       | 0        | 100%     |
| contextualized/dags/trainers.py             | 6          | 0       | 0        | 100%     |
| contextualized/easy/ContextualGAM.py        | 9          | 0       | 0        | 100%     |
| contextualized/easy/ContextualizedClassifier.py | 14    | 0       | 0        | 100%     |
| contextualized/easy/__init__.py             | 4          | 0       | 0        | 100%     |
| contextualized/easy/wrappers/__init__.py    | 1          | 0       | 0        | 100%     |
| contextualized/modules.py                   | 65         | 0       | 0        | 100%     |
| contextualized/regression/__init__.py       | 10         | 0       | 0        | 100%     |
| contextualized/regression/losses.py         | 7          | 0       | 0        | 100%     |
| contextualized/regression/metamodels.py     | 76         | 0       | 0        | 100%     |
| contextualized/regression/trainers.py       | 19         | 0       | 0        | 100%     |
| contextualized/utils.py                     | 21         | 0       | 0        | 100%     |
| Total                                       | 2639       | 333     | 0        | 87%      |
