### 2024-04-21 #6469 @cnellington
@holl- @pescap You can expect an update to the website/documentation in the next few days to better introduce common use cases and model parameters.

Regarding specific comments about naming conventions and contextualized.easy, we currently have a substantial user base and cannot rename these modules or classes, but we will explain the modules and their functionality better in our initial demos.

This naming convention is a result of our development process over the past few years. Originally, we planned to make the PyTorch native API user-facing, with the SKLearn-style easy module as a complementary tool, but the easy modules got much better reception from our user base. We've consolidated around this direction, so the easy module and related analysis utilities have become our public-facing API but the naming conventions have stuck.
