### 2024-04-22 #6469 @fabian-s
@cnellington @holl- @pescap : thanks for these updates and comments, JOSS greatly appreciates your transparency and responsiveness.

@holl- @pescap :
please remember to update the checkmarks in your checklist as issues are being solved.
if necessary, please briefly clarify which points remain open and why / what still needs to be adressed.
