+++ docs/installation.md
@@ -1,0 +1,25 @@
+# Installation
+
+## Install Contextualized
+
+````{tab-set}
+
+```{tab-item} pip
+
+`
+pip install contextualized-ml
+`
+
+```
+
+```{tab-item} source
+`
+git clone https://github.com/cnellington/Contextualized.git && pip install ./Contextualized
+`
+```
+
+````
+
+## Developer Installation
+
+To develop on Contextualized or contribute to the project, please refer to the contributing policies in the [GitHub repository](https://github.com/cnellington/Contextualized/blob/main/CONTRIBUTING.md).
