+++ docs/_toc.yml
@@ -9,6 +9,7 @@ parts:
     - file: installation
     - file: model-fitting
     - file: model-analysis
+    - file: under-the-hood
     - file: faq
   - caption: Models
     chapters:
