### 2024-04-26 4568a66
update backend note in intro
