### 2024-04-26 5e23625
remove faq with no questions or answers
