### 2024-04-27 #6469 @cnellington
Documentation has been completely updated with intro tutorials, motivations and use cases, and API reference.
https://contextualized.ml/docs/intro.html

I believe this takes care of all outstanding reviewer requests. Please let us know if you have any other comments or questions.
