### 2024-04-28 cnellington/Contextualized#230 @cnellington
@holl- I believe the documentation ordering and the descriptions for motivation and use are addressed by the recent PR above. Let us know if you have any other comments.
