### 2024-04-29 cnellington/Contextualized#230 @holl-
@cnellington Thanks for the update! I think it's a big step in the right direction.

Before I close the Issue, could you add the appropate documentation links to the `easy_regression.ipynb` notebook?
Here are my suggestions:

* Under **2. Build and fit model**, add a link to the documentation of the models / easy API.
* Link an explanation of the bootstraps in **4. Check what the individual bootstrap models learned.**
* Add a **Further Reading** section at the end with links to the advanced notebook and most essential documentation sites.
