### 2024-04-29 cnellington/Contextualized#230 @pescap
> ontexts. Neural networks are also insufficient on their own, and do not explicitly reveal a context-specific data distribution. Contextualized models recover context-specific models and generalize to new contexts.

+1 from me, as a beginner, when I started trying the tutorials and reading the documentation.

Actually, one of my first questions was: why not use a neural network with the context as features (+[shap](https://github.com/shap/shap)).
I hope this gives more importance to these kinds of questions.
