### 2024-05-02 2d798ac
add link to comparison with linear interpretability methods
