### 2024-05-07 1f02ff2
add funding statement, remove mistaken affiliation for Wesley
