+++ README.md
@@ -4,10 +4,11 @@
 ![License](https://img.shields.io/github/license/cnellington/contextualized.svg?style=flat-square)
 ![python](https://img.shields.io/badge/python-3.8%20|%203.9%20|%203.10%20|%203.11-blue)
 [![PyPI version](https://badge.fury.io/py/contextualized-ml.svg)](https://badge.fury.io/py/contextualized-ml)
-![Maintenance](https://img.shields.io/maintenance/yes/2023?style=flat-square)
+![Maintenance](https://img.shields.io/maintenance/yes/2024?style=flat-square)
 [![Downloads](https://pepy.tech/badge/contextualized-ml)](https://pepy.tech/project/contextualized-ml)
 ![pylint Score](pylint.svg)
 <a href="https://github.com/psf/black"><img alt="Code style: black" src="https://img.shields.io/badge/code%20style-black-000000.svg"></a>
+[![DOI](https://zenodo.org/badge/392447369.svg)](https://zenodo.org/doi/10.5281/zenodo.11130703)
 
 
 An SKLearn-style toolbox for estimating and analyzing models, distributions, and functions with context-specific parameters.
