### 2024-05-07 #6469 @cnellington
@fabian-s Made small changes to author affiliations and added a funding section at the end of the paper. I see by default that CC BY 4.0 is the license for the paper, which is fine, but I also want to note that the code itself is under GPL-3.0 in case there's any inconsistency.

Tagged release for Contextualized v0.2.8 in Zenodo: https://github.com/cnellington/Contextualized/releases/tag/v0.2.8-zenodo
Zenodo DOI: https://zenodo.org/doi/10.5281/zenodo.11130703
PyPI release for `contextualized-ml==0.2.8`: https://pypi.org/project/contextualized-ml/0.2.8/
