### 2024-05-08 #6469 @fabian-s
> @editorialbot recommend-accept
>
> re missing DOI for "partykit": this is a JMLR publication that does not have a DOI
>
> many thanks again to @holl- and @pescap for their reviews and congrats to @cnellington and coauthors!
