### 2024-05-08 #6469 @crvernon
:wave: I am the AEiC for the Data Science, Artificial Intelligence, and Machine Learning (DSAIS) track of JOSS submissions.  This is a very clean submission and made my job easy in this case!  Thank you! I am pushing this through to publication.
